requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dlib==20.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.1.1
flask-cors==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
numpy==2.2.6
opencv-python==4.11.0.86
pillow==11.2.1
requests==2.32.4
urllib3==2.5.0
Werkzeug==3.1.3
